Let Caladan's C++ thread suppport both copyable and move-only functions/lambdas.
Previously, Caladan's C++ thread uses std::function for its implementation.
One notable disadvantage is that std::function cannot be constructed from
a move-only lambda. A new C++ standard proposal P0288R7 is trying
to address this issue by proposing a move-only function; it is not approved
yet, but is likely to get in C++23. This commit leverages the folly::Function
which is Facebook's implementation of move-only function.