misc: set up code config file
Note that many of the source files do not strictly follow the google
style guide (for instance, whether template typename declarations are
listed on the same line as the function signature).

We could tweak this format file to minimize code changes when applying
clang-format changes. Alternatively we could make all source files
follow this minimal config in another commit.