nu: return actual data in ShardedVector::operator[]
The previous commit changed this operator to return a remote raw
pointer.

The underlying raw pointer may no longer be valid when it is
dereferenced: for instance, if the vector is resized up, prior pointers
to vector elements are no longer valid.

Returning actual data is probably what the caller wants anyway. If a
remote pointer is returned, the caller will probably dereference it,
incurring an additional round trip.